Capa de datos completada Se completaron las interfaces faltantes de los DAOs Se implementaron las interfaces faltantes de los DAOs Se agregaron clases de prueba para las implementaciones de los DAOs Se modificó el script de base de datos

+++ src/mx/unam/dgtic/modelo/dao/SolicitanteDao.java
@@ -27,14 +27,14 @@ public interface SolicitanteDao {
      * @param id
      * @return 
      */
-    public abstract SolicitanteDto eliminar(short id);
+    public abstract SolicitanteDto eliminar(int id);
     
     /**
      * Obtiene el Solicitante con el id solicitado
      * @param id
      * @return 
      */
-    public abstract SolicitanteDto obtener(short id);
+    public abstract SolicitanteDto obtener(int id);
     
     /**
      * Obtiene la lista de todos los Solicitantes almacenados
